Implement Other screen with basic layout and navigation support
Added layout for the Other screen to integrate with the app’s navigation flow.
Screen structure is prepared for future content additions and interactions.
Modified NavigationActions and made it open to use it for the tests.
Tested with a code coverage of 77% for now, other lines of code can't be covered